requirements.txt
aiohttp
discord.py
python-dotenv
aiosqlite
selenium
beautifulsoup4
pytz